backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# File Handling
Pillow==10.1.0
PyPDF2==3.0.1

# AI Integration
openai==1.3.7
google-generativeai==0.3.1

# Payment
requests==2.31.0

# Server
gunicorn==21.2.0
whitenoise==6.6.0

# Environment
python-decouple==3.8
python-dotenv==1.0.0

# Utilities
pytz==2023.3
